+++ app/services/radical_retry.rb
@@ -1,8 +1,11 @@
 class RadicalRetry
+  # do not add Timeout::Error, I think this is what is used in the rake session stuff and might conflict, dunno
+  # do not add Twilio::REST::TwilioError, else the SMS sender error handling won't work
+  # if we find we need to add the above, we can allow passing an exception to exclude to perform_request
+
   RESCUABLE_ERRORS = [Net::OpenTimeout, OpenURI::HTTPError, HTTPClient::ConnectTimeoutError, Errno::EPIPE, SocketError,
-                      OpenSSL::SSL::SSLError, Errno::ENOENT, Errno::ECONNRESET, Twilio::REST::TwilioError,
-                      Net::ReadTimeout, Errno::ECONNREFUSED, ActiveStorage::FileNotFoundError,
-                      HTTPClient::ReceiveTimeoutError].freeze
+                      OpenSSL::SSL::SSLError, Errno::ENOENT, Errno::ECONNRESET, Net::ReadTimeout, Errno::ECONNREFUSED,
+                      ActiveStorage::FileNotFoundError, HTTPClient::ReceiveTimeoutError].freeze
 
   class << self
     def perform_request(no_delay: false, retry_count: 5, additional_errors: [], &block)
